interpretability/requirements.txt
# Core dependencies
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
anthropic==0.37.0

# API dependencies for model inference
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Image processing
Pillow==10.0.0

# Web interface for human experiments
gradio==3.50.2

# Utilities
requests>=2.25.0
python-dotenv==1.0.0
retry>=0.9.2

# Statistical analysis
statsmodels==0.14.0
scikit-learn==1.3.0

# Progress bars and logging
tqdm==4.65.0

# Type hints
typing-extensions>=4.0.0 

# Testing
pytest==7.4.0 